+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/socket/SocketFactory.kt
@@ -65,9 +65,8 @@ internal class SocketFactory(
             when (connectionConf) {
                 is ConnectionConf.AnonymousConnectionConf -> "$baseWsUrl&stream-auth-type=anonymous"
                 is ConnectionConf.UserConnectionConf -> {
+                    tokenManager.ensureTokenLoaded()
                     val token = tokenManager.getToken()
-                        .takeUnless { connectionConf.isReconnection }
-                        ?: tokenManager.loadSync()
                     "$baseWsUrl&authorization=$token&stream-auth-type=jwt"
                 }
             }