+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/Mother.kt
@@ -257,7 +257,7 @@ internal object Mother {
 
     fun randomDownstreamMessageDto(
         attachments: List<AttachmentDto> = emptyList(),
-        channel: ChannelInfoDto? = null,
+        channel: ChannelInfoDto? = randomChannelInfoDto(),
         cid: String = randomString(),
         command: String? = randomString(),
         created_at: Date = randomDate(),
@@ -770,11 +770,11 @@ internal object Mother {
     fun randomDeviceDto(
         id: String = randomString(),
         pushProvider: String = randomString(),
-        providerName: String = randomString(),
+        pushProviderName: String = randomString(),
     ): DeviceDto = DeviceDto(
         id = id,
         push_provider = pushProvider,
-        provider_name = providerName,
+        push_provider_name = pushProviderName,
     )
 
     fun randomDownstreamFlagDto(