+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/channel/info/DirectChannelInfoScreen.kt
@@ -92,9 +92,14 @@ public fun DirectChannelInfoScreen(
         topBar = topBar,
     )
 
+    DirectChannelInfoScreenModal(viewModel)
+}
+
+@Composable
+private fun DirectChannelInfoScreenModal(viewModel: ChannelInfoViewModel) {
     var modal by remember { mutableStateOf<ChannelInfoViewEvent.Modal?>(null) }
 
-    LaunchedEffect(Unit) {
+    LaunchedEffect(viewModel) {
         viewModel.events.collectLatest { event ->
             if (event is ChannelInfoViewEvent.Modal) {
                 modal = event
@@ -106,6 +111,7 @@ public fun DirectChannelInfoScreen(
         modal = modal,
         isGroupChannel = false,
         onViewAction = viewModel::onViewAction,
+        onMemberViewEvent = viewModel::onMemberViewEvent,
         onDismiss = { modal = null },
     )
 }
@@ -148,11 +154,11 @@ private fun DirectChannelInfoContent(
             modifier = Modifier
                 .padding(padding)
                 .fillMaxSize(),
-            contentAlignment = if (isLoading) Alignment.Center else Alignment.TopCenter,
             isLoading = isLoading,
         ) {
             val content = state as ChannelInfoViewState.Content
             Column(
+                modifier = Modifier.matchParentSize(),
                 horizontalAlignment = Alignment.CenterHorizontally,
                 verticalArrangement = Arrangement.spacedBy(8.dp),
             ) {
@@ -177,7 +183,7 @@ private fun DirectChannelInfoContent(
                 )
                 LazyColumn {
                     items(content.options) { option ->
-                        ChannelInfoContentOption(
+                        ChannelInfoChannelOption(
                             option = option,
                             isGroupChannel = false,
                             onViewAction = onViewAction,