+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/message/internal/MessageMapper.kt
@@ -177,6 +177,7 @@ internal suspend fun ReplyMessageEntity.toModel(
             messageTextUpdatedAt = messageTextUpdatedAt,
             poll = pollId?.let { getPoll(it) },
             restrictedVisibility = restrictedVisibility,
+            channelInfo = channelInfo?.toModel(),
         )
     }
 }